is_byte_direct_readable -> is_byte_direct_ioable
AggregateType was not the correct trait as it is too lenient. Instead of
trying to build exactly what is intended i.e. a packed standard layout
type trait with trivial copy and move semantics and a 'trivial'
constructor (one that allows value initialization of its members), any
types which want to be matched should inherit from the
byte_direct_ioable_tag.

Changing this picked up some bugs too, so that's nice.