PlayerState runtime poly. -> compiletime poly.
The driving force behind this was not wanting to duplicate the event
visitation in every state, and instead allow it to be inherited somehow.
Doing so requires making PlayerState use CTRP, but then we can't have
shared_ptr<PlayerState> anymore. It's then possible to just pull the
visitation into the PlayerController anyway.

Introducing PlayerControllerImpl was necessary due to MoveAbility
needing CRTP and thus needing to include PlayerController in its header,
creating a circular dependency.