Added running
Running/walking is independent of currently implemented states, so is in
a separate state machine. Due to the (yet to be added) swim state, the
interaction between the run/walk machine and the stand/jump/sneak etc
machine is nontrivial when computing movement speeds, so the second
machine needs to know the state of the first. Moving the movement speed
outside into the PlayerController doesn't work, as the physics is driven
by the states. Hopefully this explains the (ugly) isRunning and
hasWeaponOut booleans.