Replaced shared_ptr with polymorphic_value
Some explanation is needed. The main goal of this branch is to make
Loader the default way of loading a new 'Nif' resource, after which
the manual resource loaders for Mesh and CollisionObject can load the
appropriate 'Nif' resource, which will only need to be loaded once, and
then the manual loaders can proceed completely in parallel. It initially
seemed that deep copying the BlockGraph would be necessary for thread
safety, in which case polymorphic_value (proposed for C++20) seemed a
good fit. However, each LoaderState only accessed the Block part of the
TaggedBlockGraph with const methods, so a copy is not needed afterall.

It remains to be seen if polymorphic_value will be kept, replaced by
shared_ptr again, or replaced by something differently, but the side
effects of improving const correctness and highlighting how ugly
dynamic_cast checks are will stay.