Added EspCoordinator
The problem is that multiple esp files need to be opened and read,
and with deferred records possibly simultaneously from different
threads. Moreover, the esp files have a well-defined load order that
changes the FormIds in records. This class is intended to abstract away
those issues from the resolvers and processors.