Removed engine namespace
The existence of this namespace relied on the assumption that code
would be dividable into two groups; one containing rendering and input
code, and one containing (still potentially game-specific) material that
could be used with other applications.

Clearly this assumption is incorrect. While such a separation is
well-intentioned, it is far too coarse, and also ignores the ambiguous
cases of code which depend very much on the engine, but could still be
used externally (all the Ogre extensions, for example).

The engine folder has reduced to a loose collection of 'engine-y' things
which haven't been decoupled fully into a subproject, and the namespace
serves as a functionally useless transfer of the folder structure into
code. TotW 119 and 130 are particularly relevant here.