Added dedicated NifResource
Instead of MeshLoader, CollisionObjectLoader, etc. each creating their
own BlockGraph, they access a central NifResource which does it for
them. Since Ogre's resource manager is clever enough to cache resources,
this reduces the number of calls to createBlockGraph and reduces the
amount of disk IO.

Observed load time speedup was ~5% in the current test cell.