Documented (Sub)Record, made type storage uint32_t
Originally the record type was stored as the string representation, but
the integer representation is sufficient for most purposes and can
easily generate the string representation when necessary.

The std::string specialization of Subrecord has also been removed, as
the only change was modiying size(), which could be done by modifying
SizeOf() instead.