Store DOOR, LIGH, STAT directly
Maintaining two versions of each record is impractical, and doing using
records directly doees not seem to have a noticeable negative
performance impact. Hopefully the records can be flattened in the future
so there won't be much difference in the interface anyway.

I'm sure I had an argument at one point that this was necessary for
correctly tracking save game record changes too, but I can't remember
what that was any more so it might well have been wrong. If you can
think of a good reason, let's say it was that.