Resolver 'concept' -> generic Resolver
All the current Resolvers are broken in that they do not support runtime
modification or addition of records, including from save games. Doing so
is nontrivial but generic, which provided the catalyst to make Resolver
more generic and actually work without specialization. Currently the
functionality should be sufficient to implement all Resolvers by only
specializing reifyRecord, except for those with hierarchical subgroups
or those that require deferred loading.