Split reifyRecord in two, modular REFR records
This is a big one. In brief, instead of reifyRecord producing a REFR and
a Ogre object, there needs to be an intermediate function that produces
a REFR from a base record (citeRecord) and a function that produces an
Ogre object from a REFR (reifyRecord). This is problematic as REFR alone
is insufficient to determine (at compile-time) which kind of object to
create. Thus there are now individual REFR_xxxx records that encode the
base record type in them, e.g. REFR_DOOR comes from a DOOR base record.
Since lots of REFR_xxxx share similar information, there's some (hacky)
metaprogramming to reduce code duplication, which needs cleaning up.