Fixed formatting of Doxygen
Doxygen interprets `'.'` as an empty TeX quoted empty string, followed
by a . (which ends the brief description), then a single quote, then the
start of a code span. `.` doesn't work either as the dot still ends the
brief description. I think if autobrief is on then an explicit \brief
should disable the autobrief behaviour, which (partially) solves this.