Added expression parsing, more tests to come
Attempting to sufficiently test expressions has made it clear that the
current way of checking the AST is repetitive and difficult to read,
even with helper functions. Tests should be correct by inspection, and
these require drawing diagrams to graph what they are even representing.

Clearly something must be done to make testing the grammar more pleasant
before moving on.