Added JIT support
This is highly dependent on LLVM version, and specifically requires the
7.0.0 release. The API has changed since, and much of this code will
need a rewrite when LLVM 8.0.0 is released.

The `Func` function is for example purposes only, of course. To be used
by the JIT, any external (to the JIT, so internal to the rest of the
project) function must be exported. For ELF, either all symbols in the
binary (the scripting static library only, hopefully...) need to be
exported or a list has to be given to the linker. The first method is
much easier and is used currently despite the symbol bloat because LTO
keeps the binary size down.