Fixed use-after-move in deleter
Never store a lambda capturing `this` in a class, not even indirectly!
I've made this mistake once before, you'd think I'd have learnt.

This manifested as a crash during ~Cell iff the console was opened
during gameplay, which looked like the destructors were happening out
of order (of course they weren't).

Basically the previous deleter for the std::unique_ptr<PlayerController>
captured a pointer to the GameMode in order to get a pointer to the
cell, but of course that pointer became invalidated when GameMode is
moved, which occurred when ConsoleMode was added to the mode stack and
the Mode vector was resized.