Stop Ogre from searching the Bsa files for scripts
We know the Bsa files do not contain Ogre scripts, it is pointless for
them to be searched but Ogre does not seem to provide a way of stopping
it. A patched version of Ogre is required for this, just a small if
statement in initialiseResourceGroup, it's pretty obvious.

This gives ~20% faster load times and ~20% fewer instructions on my
machine!