Fix MenuContext lifetime problems
The first (obvious) problem is that the menu uiElement is not
transferred to the context, and so any menu traits crash at runtime. The
more subtle problem is that traits rely on pointers to transfer between
the concrete and dynamic representations, but pointers to objects are
invalidated when those objects are moved. In other words, Traits and
MenuInterfaceVariant are not trivially-relocatable and so cannot just be
moved. Wrapping everything in a unique_ptr means the objects stay alive
when loadMenu finishes, but they are not moved.