Make gui and strings XML reading more consistent
Also made strings.xml optional since it is often unnecessary in tests.

So many different ways to get to an XML document; read from a Resource,
a stream, a document, several combinations of these in order, etc. Now
reading from a Resource happens as far up as possible, with xml_nodes
preferred most of the time.