MenuContext destructor deletes its Ogre::Overlay
It gets cleaned up at the end of the application, but closing the menu
needs to delete it too so that reopening the menu doesn't make a new
overlady with the same name as the old one.

An alternative approach would be to look for an existing overlay with
the correct name instead of creating one each time, giving a potential
optimization for large menus.