Cache trait values to lower update() calls
This is a bit of a hack to avoid calling update() on traits which do not
need updating, thus avoiding a lot of work in UiElements. invoke() is
still called on each trait; ideally only the traits that have changed
and their dependents would be touched, but that's considerably more
work to implement.