Add ShowMainMenu console function for testing
This---and all mode transitions occuring during console functions or
other script execution---requires a bit of trickery to ensure that the
`this` pointer of the running mode isn't invalidated immediately. The
new mode is created in a `deferredMode` member of the Application and
moved onto the top of the stack when update() finishes.