Basic interface between SoLoud and Ogre
The idea is to have an Ogre::SoundManager (analogous to an
Ogre::OverlayManager) that is used to queue sounds and play music,
then treat each mp3/wav file as an Ogre::WavResource. The SoundManager
defers to the Ogre::WavResourceManager to load the desired WavResource,
which internally contains a SoLoud::AudioSource. When a sound is played
the SoundManager instances the AudioSource into a SoLoud::Handle and
returns it wrapped in an Ogre::SoundHandle. The user accesses the
instanced sound directly through the SoundHandle instead of SoLoud's
approach of passing the handle to the Soloud singleton.

3D sounds (not yet supported) will be Ogre::MovableObject instances
backed by an Ogre::SoundHandle.