Add support for btCompoundShape
This is a bit awkward because btCompoundShape doesn't own its children.
CollisionObject can now store a vector of unique_ptrs to collision
objects as well as its main collision object, giving it ownership over
the children. A consequence is that the CollisionObjectVisitor has to
accumulate the collision objects into a vector as it parses the nif.