Use spring-based PlayerController
Instead of a capsule the height of the player, the capsule is only the
height of the upper body, but is 'attached' to the ground via a spring
with natural length equal to the length of the legs. This adds some
natural leg bending when landing from a jump and reduces lift while
going up and down slopes.