Preliminary handling of actual LAND textures
Blend maps are not currently supported because they would ideally be
generated at the same time as the rest of the LAND texture information,
but OGRE requires the layers before the terrain is loaded and the the
blend maps after.

This commit is effectively a checkpoint before I continue to attempt to
refactor the blend map generation out of the layer generation.