Support up to 9 terrain layers
This is achieved by assigning quadrants with more than 5 textures a
different material which has an extra pass of 4 additional blend layers.

Texture arrays were not a viable solution as they require all textures
to be the same size (ok until high-resolution texture mods are used)
and all the same pixel format (never ok, DXT1 and DXT3 are mixed).