Render the sun billboard above the clouds
The positions are a bit finicky because the sun texture is so wide, and
hence the billboard can clip through the cloud layer. Moving the sun
further away stops the clipping, but then its too small and making it
larger causes clipping again...

Also, the sun positioning in the sky is still wrong; it just follows
a circular arc over the player from sunrise to sunset.