Use physical-based sun position
This needs to integrate with the actual time of year to make sense, as
it currently just uses the epoch. Obviously this approach doesn't work
outside Nirn, and requires explicit calculation of the sunrise/sunset
times instead of using hardcoded ones, but I think it's cool. Will
probably make the old sun position the default and add this as an
option somehow.