Begin making the loading screen functional
There's a lot going on with this commit:
 - Remove the old KeepStrategy. It hasn't been used for a while and is
   unnecessarily complicated; we don't need to change strategies at
   runtime, at least not yet.
 - Introduce CellCache and WorldCache. These are ring buffers used to
   keep a user-definable number of cells in memory to improve loading
   times of previously visited areas.
 - Introduce CellPacket and CellRequest. These are temporary answers
   to the question of how to pass data between Modes? Stick it in a
   struct and pass it to the constructor. This is pretty ad hoc, and
   also the constructor is not the right place to be passing
   information, the enter() method is. So this needs changing.
 - Move as much of the cell and worldspace loading out of GameMode and
   into LoadingMenuMode, using the CellRequest and CellPacket to pass
   data in and out. While the loading has been moved, it all happens in
   the constructor instead of in jobs so the screen doesn't actually
   show anything yet.

In addition to the outstanding tasks listed in the above, and the
numerous TODOs, exterior cell loading within the GameMode does not take
advantage of the CellCache, and cached but unused cells are still
present and unhidden in the scene manager. Consequently exterior cells
have a double loading bug where moving into a cached cell---which still
has all its nodes attached to the world's scene manager---added again.