Hide 'unloaded' cached cells
A cell that is no longer in the player's near neighbourhood but remains
in the cache has its nodes hidden and its physics objects detached
from the physics world. These are unhidden and reattached if the cell
comes back into the near neighbourhood from the cache.

It was also necessary to make the lifetime of the WorldCache greater
than the lifeteime of the CellCache because the destruction of an
exterior cell requires its scene manager and physics world---owned by
its parent World---to still be alive.

This does *not* fix all the lifetime issues with cached ExteriorCells;
if their parent World is removed from the cache, then all the cached
cells that it owns should be removed first.