Fix object positions and rotations
This requires some cleanup before merging. In particular, the decision
as to when transformations are incorporated into SceneNodes vs when
they are baked into the collision objects/meshes is unclear. It is also
quite cumbersome to even apply; mUndoRootTransform and the additional
start vertex overloads are awkward and feel hacked-on (because they
are...)