Load persistent records in exterior cells
I don't like how I've implemented this, but it is not too unreasonable
to add an insertReferenceRecord to CellResolver because (in the future)
reference records can be created dynamically.

I considered adding the persistent records explicitly when the cell is
loaded, without adding them to the cell's references list, but that
means duplicating the responsibility of reification in the cell.
Moreover, the reference list of an exterior cell would contain a
fundamentally different set of records to an equivalent interior cell.

It is still unfortunate that the persistent record insertion is
decoupled from the actual loading of the cell. Perhaps
CellResolver::load() should be given an additional
PersistentReferenceLocator parameter instead?