Fix WorldCache-CellCache inconsistency bug
When changing worlds after filling the WorldCache, the oldest World
is pushed from the cache. If the CellCache still contains ExteriorCells
belonging to that World, then when they destruct they will attempt to
remove themselves from a World that no longer exists.

The fix taken here is to join the WorldCache into the CellCache, where
really it should have been all along, and set things up so that when
a World is pushed out of the cache it removes all its cached child
Cells too. Currently that's done using a move-only wrapper around the
WorldPtr that calls an invalidate() method in its destructor and
move-assignment operator.

It would be nice if we could make use of a custom deleter, since we're
already using shared_ptr, but because the deleter is stored in the
control block it can't be changed for a single shared_ptr (the copy
of the one given to push_back() that is stored, in particular).