Fix getLandId() logic
This was implemented with the understanding that the LAND record of
a child cell should take precedence over the LAND record of its parent,
but actually that is not true. Even though some cells in a child
worldspace have LAND records, these should never be used and the LAND
record of the parent always takes precedence. This can be confirmed in
the CS by creating a worldspace and child worldspace, modifying the
terrain in the child, and observing that the change occurs in the
parent.