Simulate Havok collision layers with Bullet groups
Oblivion uses collision layers instead of collision groups. Masks are
unique for each layer, instead of specific to each object, and each
object can only be in one layer at a time.

This commit also changes the direction of the bullet directory, using
it now as engine-specific bullet integration code, instead of trying to
be a generic wrapper. The Ogrebullet module already does most of the
wrapping we need, and with the inclusion of collision layers a place
was needed to put common physics-oriented functions.