Begin documenting the animation system
While being stuck figuring out how to implement animations embedded in
nif files I realised that I had forgotten parts of how OGRE's animation
system worked and not documented the reasoning leading up to the changes
made to OGRE's animation system, only why the changes were made (and
that they were probably not the best way to fix the encountered
problems). This commit begins to document in detail the steps necessary
to use hardware skinned skeletal animation in OGRE and the conflicts
between OGRE's animation system and the NIF format's system (or at least
my interpretation of it with a lot of help from NifSkope), both as a
resource to refer to if I forget how the system works again (or someone
else wants to know) and as a way of reasoning out how I should fix my
current attempt at integrating the two systems.