Support Ogre::Mesh and oo::Mesh simultaneously
oo::Mesh and oo::Entity are now named as such, instead of replacing
the defaults, mainly because Ogre assumes it's own classes are in use
in some places and breaks if they aren't. In particular, the
SceneManager's sky box methods are nonvirtual so cannot be overridden to
use oo::Mesh. Moreover, InstancedMesh would have to be reimplemented too
and I can't be bothered just for the water planes.