Attempt to forward render transparent objects
This doesn't work, and I'm not completely sure why. Firstly, OGRE's
default behaviour with material schemes is to render an object with its
first technique if it doesn't have a technique for the active scheme;
this is surprising (to me at least) since it means that you cannot
prevent the rendering of an object by changing schemes. I may have
misunderstood the tutorial and it turns out that you have to give every
material a NoGBuffer technique, but just make the deferred materials'
ones empty.

Modifying OGRE to return nullptr when it doesn't have a technique for
the active scheme does not fix the problem; it fixes the UB from
rendering the deferred items in the forward stage, but then the forward
objects don't render at all. Specifying the material_scheme explicitly
for the GBuffer objects to "GBuffer" prevents them from rendering
either, but works fine if the material_scheme is given for the target
instead of just the pass. This suggests that specifying the
material_scheme on a per-pass level has some issues. Possibly caused by
my modification of course.