Some progress with deferred rendering
Continued following OGRE's deferred rendering tutorial and added light
meshes. These need to be stored somewhere and since they're part of the
scene it makes sense to store them with the lights in the SceneManager.
This means constructing a custom SceneManager (which opens the way for
other optimizations unrelated to rendering).

Unsurprisingly, the rendering still has problems. Firstly, the gamma
correction is wrong; all the lighting must be done in linear space,
which is now accumulated in the final framebuffer. This means that the
framebuffer must be linear with a gamma correction pass being run at
the end instead of at the end of each mesh's fragment shader. There is
no such pass yet (see below) so gamma has been mostly ignored.

Secondly, the ambient light is not rendered if the target is a render
texture. The other lights render fine, just not the ambient light, and
everything renders ok to the screen, just not to a texture. This doesn't
just include when target_output is replaced by target rt0 (for example),
it includes when any compositor pass is added after the compositor pass
that renders the lights. Adding a gamma correction pass for example
therefore prevents the ambient lighting from appearing. I've spent hours
on this trying to work out what's going on, and I've run out of ideas.
Nsight just says that the fullscreen ambient light quad draw call draws
no fragments, regardless of what I set its fragment shader to output,
and whether it considers a depth buffer or not.