Add initial InteriorSceneManager
With the default SceneManager there are many draw calls per frame that
don't result in any fragments, so culling could be a lot better; most
of the frame time is spent finding objects to render and pushing them
to the render queue.

InteriorSceneManager is an initial attempt at using an octree to improve
occlusion culling for interior cells. Because it uses an octree with
fixed bounds, a different approach will be needed for exterior cells.

This is currently unoptimized and isn't any faster than the default
scene manager, though hopefully it can be made quicker.