Begin abstracting NPC animations into anim groups
Instead of specifying animations directly, NPCs are told to perform
an AnimGroup which encapsulates a related set of animations. The
particular animation should be chosen depending on the NPCs state, such
as whether they have a weapon out, and if so, what kind.
See the CSWiki entry on AnimGroups for more information.

These functions all need to be moved somewhere else but I haven't
decided where yet.