Wrap player and NPCs into Character class
NPCs and the player are fundamentally very similar from a physics and
gameplay point of view, differing mostly in the way they are controlled.
Because the CharacterController operates using events and not direct
input, it can just as easily obtain its input events from an AI system.
In that way CharacterController can be used for both player and NPCs.
The idea is that the Character class will hold both the controller and
any additional information previously used by the NPC resolver, such as
base records, entities, and so on.

There are still outstanding issues here, mostly that the NPCs each have
their own camera, and that the skeleton and collision capsules are
joined together in an ad-hoc fashion.