Use specular textures
It turns out that the alpha component of the normal map is used as
the specular intensity, if the alpha channel is present. This is hard to
check, so a compromise is made where alpha values of 255 are mapped to
zero, but every other value is used as a specular intensity. Because the
terrain has separated normal and diffuse passes, the alpha component
of the normal GBuffer now also stores the specular, whereas the old
albedoSpec GBuffer continues to store the material shininess but is now
just called albedo.