Parent Nif root of skeleton to capsule
The camera is now parented to the skeleton, though just the root node
and not the head node, so no true first person just yet. Because the
body node controls the physics, the skeleton has to be a child of it,
and hence the root node is not actually the root; the body is.
moveTo() therefore has to adjust for the body/root difference, and the
feet clip through the floor when landing. I think this is fixable by
applying the opposite of the body node spring displacement to the root
scene node.

The NPC and player models do not rotate correctly yet.