Save nif node names, fix nif light placement
Mesh lights have NiNodes called 'AttachLight' which the light should be
attached to, it should not be attached to the root. Naming the nodes is
a bit awkward because Ogre doesn't allow duplicate SceneNode names.

It would be useful to have the RefId of the object being loaded to
guarantee uniqueness, but since that's not available currently (though I
see no reason why it couldn't be passed through) an increasing 64-bit
counter is used instead.