Make CharacterControllerImpl public vars private
Some of these getter/'setter' pairs are questionable e.g. getPitch,
which returns a reference instead of having a setter; this is just
because the common idiom is to increment not set. Moreover, when
animation accumulation is added (soon I hope) much of this will need to
be removed or refactored out.