Add fog
Unfortunately this was not as simple as updating the PostMaterial
uniforms each frame (say in DeferredLightPass), because the compositor
uses a copy of its given material instead of the original. A listener
has to be created and linked to (owned by, to prevent ownership issues)
the scene manager, which then takes the scene's fog parameters and sets
the uniforms on the compositor's copy of the material.