Remove Abseil
Working with Abseil so far has been nice from a programming point of
view but ultimately it isn't a good fit for this project. Mostly this is
due to issues involving CMake, which when Boost is also present and can
do a lot of what Abseil does, just aren't worth the effort.

For a while Abseil has been fixed at commit eab2078 on my machine due to
its refusal to export its targets when used via add_subdirectory.
(At the time I thought this was not worth committing but now I think I
was wrong and definitely should have committed it.)

Since the FetchContent workflow relies upon this, and projects should
export their targets, it becomes impossible to use Abseil in any of the
OpenOblivion* subprojects. Without FetchContent 'living at head' is
considerably more effort (no automatic updates) so either way updates
end up not being followed.

Even if it is possible to integrate Abseil into CMake in a way that
allows for automatic updates without explicit update commits and that
supports subprojects, when Boost can take its place in most cases
it doesn't seem worth looking for such a solution.