Refactor LoadingMenu cell loading
The cell loading is some of the more complicated code in this project,
so hopefully this makes it clearer to see what's going on (or what
should be going on, at least). Much of the process is shared with
ExteriorManager, but it has been hard to share the code due to the
different assumptions on the starting state before a load. My hope is
that with the right structure, a lot more of the code can be
deduplicated.