Load distant LOD terrain
Distant LOD meshes and textures for the terrain are baked by the artist
ahead of time and stored as (almost) normal mesh and texture files.
Each baked mesh represents a 32x32 chunk of cells and is present at all
times. This includes the chunk that intersects the near neighbourhood,
currently causing clipping issues.