Begin replacement of CharacterController
The current CharacterController is moved into Character, as are the
member variables of CharacterControllerImpl. A CharacterMediator is
passed to the states instead, working something like the Attorney-Client
pattern. The old character_controller_impl.hpp files etc. are kept
around for reference and will be removed later.

All the physics code has been removed in preparation for a new character
controller system that integrates better with the skeleton, supports
animation properly, and hopefully doesn't fall through steep slopes.