Add FetchContent for Bullet, CMake 3.12 -> 3.13
The default is to use FetchContent, since this makes things much easier
for the user, but find_package can be used instead by setting the
option OO_USE_EXTERNAL_BULLET to ON. This allows patched versions of
Bullet installed locally to be used instead, e.g. by modifying
Bullet_ROOT or CMAKE_PREFIX_PATH.

Quite a bit of work is needed to persuade Bullet to work properly with
targets; target_include_directories and target_link_libraries need to be
set for each target and the targets need to be exported. Lots of options
also need to be set to disable the parts of Bullet that we don't need.