Move BsaArchiveFactory to Ogre subproject
Originally the Ogre subproject was seen as containing components that
may be useful elsewhere to projects using OGRE, but fnt_loader.hpp and
tex_image_codec.hpp already go against that. Instead we'll fully commit
to its interpretation as utilities specific to OGRE, that wouldn't be
needed if OGRE wasn't being used.