Change {cite,reify}Record() customization point
Instead of the user specializing citeRecord() and reifyRecord(), they
should should instead specialize the CiteRecordImpl and ReifyRecordImpl
class templates which replace and extend CiteRecordTrait and
ReifyRecordTrait.

This fixes some akwardness with overloading reifyRecord for CELL and
WRLD stemming from using specialization of a function template as a
customization point. Thanks to Herb Sutter's advice in 'Why Not
Specialize Function Templates' in particular, which I am many years late
on.