Create util subproject
This subproject defines a header-only library containing code that is
useful in many different locations and that does not bring in any
dependencies. Specifically it is intended for generic code that could
be made use of outside the project, if any of it was actually novel
enough. settings.hpp is also there too, since it needs to go somewhere.