Be more specific about boost adaptors
The generic adaptors header includes the tokenized adaptor, which in
turn uses boost::regex. On GCC and Clang this isn't a problem because
we never use the regex functionality, but MSVC complains that it can't
find boost::regex when linking unless we explicitly link to it.

Maybe MSVC is trying to do some sophisticated dependency detection? It
shouldn't be an issue because we never call any functions in
boost::regex. The only mention I could find of needing to link to
boost::regex at all was in the compiled object files of the two
modified cpp files, so I'm not sure exactly what's going on here.

Anyway, the smaller header should be used even if MSVC didn't have this
behaviour, so thanks MSVC, I guess?