Attempt to centralise windows.h fixes
This approach of including a header is a little less ad-hoc than just
undefining macros that windows.h has claimed, and is a little more
robust. It currently breaks encapsulation of the cmake subprojects
though, since the header is only available due to it happening to be
in the include path, and not because it was explicitly requested.