Fix possible UB in verOf
I'm not 100% certain what is going on here; only MSVC in debug mode
refuses to compile this, release mode doesn't complain and neither do
GCC or Clang in any configuration.

It appears that the problem is that because the check for a new chunk is
done before the bounds check, at the end of the loop one-past-the-end is
read, and that's UB. Since this is being done in a constexpr context,
compilation stops with a (rather unhelpful) 'not a constant-expression'.

Interestingly, MSVC is perfectly fine if instead of evaluating verOf
in the template argument list directly, the result is first stored in a
static constexpr inline class variable and then that is passed as an
argument.

I think if str was a const char * this would be fine since it would be
given a null-terminated string-literal (unless it's UB to dereference
the null-terminator?), but because the argument is a std::string_view,
which may not be null-terminated, this is an out-of-bounds access.

Anyway, the conditions are definitely the wrong way round whatever's
going on.