Use std::size_t for Record and Subrecord sizes
The data type used for serialization of the size is an implementation
detail, and should not be part of the interface; any conversion is now
just done in the IO functions, instead of happening implicitly through
return types. There's no checking that types actually fit within the
size bounds, but then there wasn't before either.