Remove BsaReader::FolderAccessor
The operations to open a FileData and get the uncompressed size of a
file are now member functions of BsaReader. It would be nice if they
could be member functions of a FileView, but doing that requires
storing a pointer to the BsaReader and propagating it through
FolderView, FolderIterator, FileIterator, and finally to FileView, which
seems a bit much.